move sdk extensions to commons
