+++ app/src/main/java/org/dhis2/utils/granularsync/GranularSyncRepository.kt
@@ -3,6 +3,36 @@ package org.dhis2.utils.granularsync
 import io.reactivex.Single
 import java.util.Locale
 import org.dhis2.R
+import org.dhis2.commons.bindings.categoryOptionCombo
+import org.dhis2.commons.bindings.countDataValueConflicts
+import org.dhis2.commons.bindings.countEventImportConflicts
+import org.dhis2.commons.bindings.countTeiImportConflicts
+import org.dhis2.commons.bindings.dataElement
+import org.dhis2.commons.bindings.dataSetInstanceSummaries
+import org.dhis2.commons.bindings.dataSetInstancesBy
+import org.dhis2.commons.bindings.dataSetSummaryBy
+import org.dhis2.commons.bindings.dataValueConflicts
+import org.dhis2.commons.bindings.dataValueConflictsBy
+import org.dhis2.commons.bindings.enrollment
+import org.dhis2.commons.bindings.enrollmentImportConflicts
+import org.dhis2.commons.bindings.enrollmentInProgram
+import org.dhis2.commons.bindings.event
+import org.dhis2.commons.bindings.eventImportConflictsBy
+import org.dhis2.commons.bindings.eventsBy
+import org.dhis2.commons.bindings.observeDataSetInstancesBy
+import org.dhis2.commons.bindings.observeEvent
+import org.dhis2.commons.bindings.observeProgram
+import org.dhis2.commons.bindings.observeTei
+import org.dhis2.commons.bindings.organisationUnit
+import org.dhis2.commons.bindings.period
+import org.dhis2.commons.bindings.program
+import org.dhis2.commons.bindings.programs
+import org.dhis2.commons.bindings.tei
+import org.dhis2.commons.bindings.teiAttribute
+import org.dhis2.commons.bindings.teiImportConflictsBy
+import org.dhis2.commons.bindings.teiMainAttributes
+import org.dhis2.commons.bindings.teisBy
+import org.dhis2.commons.bindings.trackedEntityType
 import org.dhis2.commons.date.toUi
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.resources.ResourceManager
@@ -10,36 +40,6 @@ import org.dhis2.commons.sync.ConflictType
 import org.dhis2.commons.sync.SyncContext
 import org.dhis2.data.dhislogic.DhisPeriodUtils
 import org.dhis2.data.dhislogic.DhisProgramUtils
-import org.dhis2.metadata.usecases.sdkextensions.categoryOptionCombo
-import org.dhis2.metadata.usecases.sdkextensions.countDataValueConflicts
-import org.dhis2.metadata.usecases.sdkextensions.countEventImportConflicts
-import org.dhis2.metadata.usecases.sdkextensions.countTeiImportConflicts
-import org.dhis2.metadata.usecases.sdkextensions.dataElement
-import org.dhis2.metadata.usecases.sdkextensions.dataSetInstanceSummaries
-import org.dhis2.metadata.usecases.sdkextensions.dataSetInstancesBy
-import org.dhis2.metadata.usecases.sdkextensions.dataSetSummaryBy
-import org.dhis2.metadata.usecases.sdkextensions.dataValueConflicts
-import org.dhis2.metadata.usecases.sdkextensions.dataValueConflictsBy
-import org.dhis2.metadata.usecases.sdkextensions.enrollment
-import org.dhis2.metadata.usecases.sdkextensions.enrollmentImportConflicts
-import org.dhis2.metadata.usecases.sdkextensions.enrollmentInProgram
-import org.dhis2.metadata.usecases.sdkextensions.event
-import org.dhis2.metadata.usecases.sdkextensions.eventImportConflictsBy
-import org.dhis2.metadata.usecases.sdkextensions.eventsBy
-import org.dhis2.metadata.usecases.sdkextensions.observeDataSetInstancesBy
-import org.dhis2.metadata.usecases.sdkextensions.observeEvent
-import org.dhis2.metadata.usecases.sdkextensions.observeProgram
-import org.dhis2.metadata.usecases.sdkextensions.observeTei
-import org.dhis2.metadata.usecases.sdkextensions.organisationUnit
-import org.dhis2.metadata.usecases.sdkextensions.period
-import org.dhis2.metadata.usecases.sdkextensions.program
-import org.dhis2.metadata.usecases.sdkextensions.programs
-import org.dhis2.metadata.usecases.sdkextensions.tei
-import org.dhis2.metadata.usecases.sdkextensions.teiAttribute
-import org.dhis2.metadata.usecases.sdkextensions.teiImportConflictsBy
-import org.dhis2.metadata.usecases.sdkextensions.teiMainAttributes
-import org.dhis2.metadata.usecases.sdkextensions.teisBy
-import org.dhis2.metadata.usecases.sdkextensions.trackedEntityType
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.common.State
 import org.hisp.dhis.android.core.dataset.DataSetInstance
