+++ commons/src/main/java/org/dhis2/commons/bindings/SdkExtensions.kt
@@ -1,4 +1,4 @@
-package org.dhis2.metadata.usecases.sdkextensions
+package org.dhis2.commons.bindings
 
 import io.reactivex.Single
 import org.hisp.dhis.android.core.D2
